users route pulled into user routing file
